feat: add rendering of chess board
Added rendering of the chess board.
Added a ChessSquare component to render each square
Added a hook to get square objects to render the board with.
The hook takes position and flip arguments to render the pieces
and handles the flipping of the board (reversal)

Still working on functionality for the rendering.